refactor(test-harness): one report-emission tail for Foundry and Hardhat
The two runners ended in line-for-line identical dir-create + benchmark
JSON + XLSX sequences (identical modulo suite strings — this branch had
already extended both copies in lockstep once). write_reports owns the
tail, keyed by SuiteKind; the comparisons filter stays per-runner with
its typed config.
