+++ solx-dev/src/test/mod.rs
@@ -20,6 +20,47 @@ pub(crate) fn toolchain_name(compiler_name: &str, codegen: &str) -> String {
 }
 
 ///
+/// Writes a suite's benchmark JSON and XLSX report into the output directory
+/// — one emission tail shared by the Foundry and Hardhat runners so the two
+/// cannot drift.
+///
+pub(crate) fn write_reports(
+    benchmark: solx_benchmark_converter::Benchmark,
+    comparisons: Vec<solx_benchmark_converter::OutputComparison>,
+    output_directory: std::path::PathBuf,
+    kind: solx_benchmark_converter::SuiteKind,
+) -> anyhow::Result<()> {
+    use colored::Colorize;
+
+    std::fs::create_dir_all(output_directory.as_path()).map_err(|error| {
+        anyhow::anyhow!(
+            "{} {} output reports directory {output_directory:?}: {error}",
+            "Creating".bright_green().bold(),
+            kind.label(),
+        )
+    })?;
+    let base_path = crate::utils::absolute_path(output_directory)?;
+
+    write_benchmark_json(&benchmark, base_path.as_path(), kind.benchmark_file())?;
+
+    let output: solx_benchmark_converter::Output = (
+        benchmark,
+        comparisons,
+        solx_benchmark_converter::OutputFormat::Xlsx,
+    )
+        .try_into()?;
+    let mut output_path = base_path;
+    output_path.push(kind.report_file());
+    eprintln!(
+        "{} the spreadsheet report to {}",
+        solx_utils::cargo_status_ok("Writing"),
+        output_path.to_string_lossy().bright_white().bold()
+    );
+    output.write_to_file(output_path)?;
+    Ok(())
+}
+
+///
 /// Emits the merged benchmark JSON next to the XLSX so the workflow's summary
 /// step can render the PR comment from the native data model, in the
 /// converter's own JSON output format.
